requirements.txt
faiss-cpu
ollama
PyPDF2
langchain
sentence-transformers
